+++ spotifywebapipython/models/spotifyconnectdevices.py
@@ -96,6 +96,63 @@ class SpotifyConnectDevices():
         return 0
     
 
+    def RemoveDevice(
+        self,
+        deviceId:str,
+        dynamicDeviceOnly:bool=False
+        ) -> None:
+        """
+        Removes existing device entry(s) from the `Items` collection by device id value.
+        
+        Args:
+            deviceId (str):
+                Device id to remove.
+            dynamicDeviceOnly (str):
+                True to only remove the device if it's a dynamic device entry;
+                otherwise, False to remove the device id regardless.  
+                Default is False.
+
+        It's possible to have multiple devices with the same id (e.g. both a dynamic and a
+        zeroconf registered device).
+
+        Dynamic devices are Spotify Connect devices that are not found in Zeroconf discovery
+        process, but still exist in the player device list.  These are usually Spotify Player
+        client devices (e.g. mobile / web / desktop players) that utilize temporary device id's.
+        """
+        # validations.
+        if (not isinstance(dynamicDeviceOnly,bool)):
+            dynamicDeviceOnly = False
+
+        # remove device from devices collection;
+        # the reversed() function creates an iterator that traverses the list in reverse order. 
+        # this ensures that removing an element doesn't affect the indices of the subsequent 
+        # elements we're going to iterate over.
+        deviceIdCompare:str = "" + deviceId.lower()
+        spDevice:SpotifyConnectDevice
+        for idx in reversed(range(len(self._Items))):
+
+            spDevice = self._Items[idx]
+
+            # did we find a match by device id?
+            if (deviceIdCompare == spDevice.Id.lower()):
+
+                isFound:bool = True
+
+                # get dynamic device status.
+                isDynamic:bool = False
+                if (spDevice.DiscoveryResult is not None):
+                   isDynamic = spDevice.DiscoveryResult.IsDynamicDevice
+
+                # are we checking for dynamic device? if so, and it's NOT dynamic, then don't delete it!
+                if (dynamicDeviceOnly) and (not isDynamic):
+                    isFound = False
+
+                # are we deleting this device?
+                if (isFound):
+                    _logsi.LogVerbose("Removing device \"%s\" (%s) from Spotify Connect Devices collection (IsDynamicDevice=%s)" % (spDevice.Name, spDevice.Id, str(isDynamic)))
+                    self._Items.pop(idx)
+
+
     def AddDynamicDevice(
         self, 
         device:Device,